Add fixes from testing to dev

+++ src/AI/AIController.h
@@ -54,7 +54,7 @@ private:
     AI ai_;
     ai_state_t state_;
     action_t planned_action_;
-    float prev_transition_timestamp;
+    double prev_transition_timestamp;
 
 public:
