+++ src/AI/Drone.cpp
@@ -22,14 +22,14 @@ bool Drone::update(observation_t observation){
 }
 
 
-void Drone::wait(float time){
+void Drone::wait(double time){
 	//NOT IMPLEMENTED YET
 }
 
-float Drone::getDistanceToPoint(point_t point){
+double Drone::getDistanceToPoint(point_t point){
 
-	float x_Distance = point.x - this->position.x;
-	float y_Distance = point.y - this->position.y;
+	double x_Distance = point.x - this->position.x;
+	double y_Distance = point.y - this->position.y;
 
 	return sqrt(pow(x_Distance,2) + pow(y_Distance,2));
 }
@@ -38,12 +38,12 @@ float Drone::getDistanceToPoint(point_t point){
 point_t Drone::getInterceptPoint(Robot robot) {
 	
 	//Drone* drone = new Drone(); 
-	float time_Until_Turn = 20 - robot.getTimeAfterTurn();
-	float robot_Ori = robot.plank.getAngle();
+	double time_Until_Turn = 20 - robot.getTimeAfterTurn();
+	double robot_Ori = robot.plank.getAngle();
 	point_t robot_Pos = robot.getPosition();
-	float x_d = this->position.x;
-	float y_d = this->position.y;
-	float time_Since_Turn = robot.getTimeAfterTurn();
+	double x_d = this->position.x;
+	double y_d = this->position.y;
+	double time_Since_Turn = robot.getTimeAfterTurn();
 
 	if(time_Since_Turn < 2) {
 		robot_Ori = robot_Ori - (MATH_PI/2)*(time_Since_Turn) + MATH_PI;
@@ -54,46 +54,46 @@ point_t Drone::getInterceptPoint(Robot robot) {
 	}
 	
 	//Math to calculate if direct
-	float a = robot_Pos.x; float b = robot.getSpeed(); float c = robot_Ori; float d = robot_Pos.y; float e = x_d; float f = y_d; float g = this->speed;
-	float ta =(-sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
-	float tb = (sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
+	double a = robot_Pos.x; double b = robot.getSpeed(); double c = robot_Ori; double d = robot_Pos.y; double e = x_d; double f = y_d; double g = this->speed;
+	double ta =(-sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
+	double tb = (sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
 
-	float t1 = (std::max)(ta, tb);
-	float t2 = 0;
+	double t1 = (std::max)(ta, tb);
+	double t2 = 0;
 
-	float x_bf = 0;
-	float y_bf = 0;
+	double x_bf = 0;
+	double y_bf = 0;
 
 	if(t1 > (time_Until_Turn+2))
 	{
 		t1 = time_Until_Turn;
-		float x_b1 = robot_Pos.x +time_Until_Turn*robot.getSpeed()*cos(robot_Ori);
-		float y_b1 = robot_Pos.y +time_Until_Turn*robot.getSpeed()*sin(robot_Ori);
-		float angleDrone1 = atan2(y_b1-y_d, x_b1-x_d);
+		double x_b1 = robot_Pos.x +time_Until_Turn*robot.getSpeed()*cos(robot_Ori);
+		double y_b1 = robot_Pos.y +time_Until_Turn*robot.getSpeed()*sin(robot_Ori);
+		double angleDrone1 = atan2(y_b1-y_d, x_b1-x_d);
 
-		float a = x_b1; float b = robot.getSpeed(); float c = robot_Ori+MATH_PI; float d = y_b1; float e = x_d + (time_Until_Turn+2)*this->speed*cos(angleDrone1); float f = x_d + (time_Until_Turn+2)*this->speed*sin(angleDrone1); float g = this->speed;
+		double a = x_b1; double b = robot.getSpeed(); double c = robot_Ori+MATH_PI; double d = y_b1; double e = x_d + (time_Until_Turn+2)*this->speed*cos(angleDrone1); double f = x_d + (time_Until_Turn+2)*this->speed*sin(angleDrone1); double g = this->speed;
 		ta =(-sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
 		tb = (sqrt(pow(b,2)*pow(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c),2) - 4*(-pow(a,2) + 2*a*e - pow(d,2) + 2*d*f - pow(e,2) - pow(f,2))*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2))) - b*(-2*a*cos(c) - 2*d*sin(c) + 2*e*cos(c) + 2*f*sin(c)))/(2*(-pow(b,2)*pow(sin(c),2) - pow(b,2)*pow(cos(c),2) + pow(g,2)));
 		t2 = (std::max)(ta, tb);
 
-		float x_d1 = e;
-		float y_d1 = f;
+		double x_d1 = e;
+		double y_d1 = f;
 		
 		x_bf = x_b1+t2*robot.getSpeed()*cos(robot_Ori+MATH_PI);
 		y_bf = y_b1+t2*robot.getSpeed()*sin(robot_Ori+MATH_PI);
 
-		float angleDrone2 = atan2(y_bf-y_d1, x_bf-x_d1);
+		double angleDrone2 = atan2(y_bf-y_d1, x_bf-x_d1);
 	}
 	else
 	{
 		x_bf = robot_Pos.x+t1*robot.getSpeed()*cos(robot_Ori);
 		y_bf = robot_Pos.y+t1*robot.getSpeed()*sin(robot_Ori);
-		float angleDrone1 = atan2(y_bf-y_d, x_bf-x_d);
+		double angleDrone1 = atan2(y_bf-y_d, x_bf-x_d);
 	}
 	point_t intersection;
 	intersection.x = x_bf;
 	intersection.y = y_bf;
-	float t = t1+t2;
+	double t = t1+t2;
 	intersection.travel_Time = t;
 
 	return intersection;
