+++ src/AI/Observation.cpp
@@ -42,7 +42,7 @@ std::array<Robot,4> Observation::getObstacles(){
 	return this->obstacles;
 }
 
-float Observation::getTimeStamp(){
+double Observation::getTimeStamp(){
 	return this->time_Stamp;
 }
 
@@ -51,13 +51,13 @@ bool Observation::anyRobotsVisible(){
 }
 
 
-bool Observation::update(observation_t observation, float elapsed_time){
+bool Observation::update(observation_t observation, double elapsed_time){
 	bool drone_Updated = updateDrone(observation, elapsed_time);
 	bool robot_Updated = updateRobot(observation, elapsed_time);
 	return (drone_Updated and robot_Updated);
 }
 
-bool Observation::updateDrone(observation_t observation, float elapsed_time){
+bool Observation::updateDrone(observation_t observation, double elapsed_time){
 	this->time_Stamp = elapsed_time;
 	this->drone.update(observation);
 	return true;
@@ -67,7 +67,7 @@ void Observation::updateInteraction(int index) {
 	this->robots[index].setInteractedWithTrue();
 }
 
-bool Observation::updateRobot(observation_t observation, float elapsed_time){
+bool Observation::updateRobot(observation_t observation, double elapsed_time){
 	bool any_robots_visible = false;
 	point_t position = point_zero;
 	this->time_Stamp = elapsed_time;
