+++ src/AI/Observation.h
@@ -5,7 +5,7 @@
 
 class Observation{
 private:
-	float time_Stamp;
+	double time_Stamp;
 	std::array<Robot,10> robots;
 	std::array<Robot,4> obstacles;
 	Drone drone;
@@ -22,12 +22,12 @@ public:
 	Robot getObstacle(int index);
 	std::array<Robot,4> getObstacles();
 
-	float getTimeStamp();
+	double getTimeStamp();
 
 	bool anyRobotsVisible();
 
-	bool update(observation_t observation, float elapsed_time);
-	bool updateDrone(observation_t observation, float elapsed_time);
-	bool updateRobot(observation_t observation, float elapsed_time);
+	bool update(observation_t observation, double elapsed_time);
+	bool updateDrone(observation_t observation, double elapsed_time);
+	bool updateRobot(observation_t observation, double elapsed_time);
 	void updateInteraction(int index);
 };
